[IntegrationTests] Replace size_t with unsigned
size_t is not defined when compiling with Polygeist, so all references
to it have been replaced by unsigned.

+++ integration-test/src/bicg/bicg.c
@@ -8,14 +8,13 @@
 //===----------------------------------------------------------------------===//
 
 #include "bicg.h"
-#include <cstddef>
 
 int bicg(int a[N][N], int s[N], int q[N], int p[N], int r[N]) {
   int tmp = 0;
-  for (size_t i = 0; i < N; i++) {
+  for (unsigned i = 0; i < N; i++) {
     tmp = q[i];
 
-    for (size_t j = 0; j < N; j++) {
+    for (unsigned j = 0; j < N; j++) {
       int val = a[i][j];
       s[j] = s[j] + r[i] * val;
       tmp += val * p[j];
